+++ src/services/permissions/types.d.ts
@@ -3,4 +3,7 @@ export interface UserPermissions {
   isAdmin: boolean // Whether the user is an admin
   isModerator: boolean // Whether the user is a moderator
   isABasicMember: boolean // Whether the user has basic membership
+  hasAllContentAccess: boolean
+  isForumModerator: boolean
+  isCommentModerator: boolean
 }